refactor(cli): split pay policy into its own module
Addresses the round-2 structural nit on #115: pickAccept and enforceMaxAmount
were exported from pay.ts only so the tests could reach them, leaving them
part of that file's public shape by convention rather than by structure.

Move the two pure helpers to src/pay/policy.ts and promote src/pay.ts to
src/pay/index.ts. Callers (src/index.ts, src/bin.ts) now import the public
pay() surface from ./pay/index.js; the policy helpers are importable only
via ./pay/policy.js, which is not re-exported from the package root.

Tighten vitest coverage exclude from src/**/index.ts to src/index.ts +
src/bin.ts so the real logic in src/pay/index.ts and src/signers/index.ts
stays measured.

Tests: 35/35 still passing.
