+++ packages/cli/src/index.ts
@@ -7,7 +7,7 @@ export {
   SettlementError,
   SignatureRejectedError,
 } from './errors.js'
-export type { PayFlags, PayResult } from './pay.js'
-export { pay } from './pay.js'
+export type { PayFlags, PayResult } from './pay/index.js'
+export { pay } from './pay/index.js'
 export { resolveSigner, SignerResolutionError } from './signers/index.js'
 export type { ResolvedSigner, SignerFlags, SignerKind } from './types.js'
