+++ packages/cli/tests/helpers.test.ts
@@ -1,7 +1,7 @@
 import type { PaymentRequirements } from '@x402/core/types'
 import { describe, expect, it } from 'vitest'
 import { Malformed402Error, MaxAmountExceededError } from '../src/errors.js'
-import { enforceMaxAmount, pickAccept } from '../src/pay.js'
+import { enforceMaxAmount, pickAccept } from '../src/pay/policy.js'
 
 const USDC = '0x036CbD53842c5426634e7929541eC2318f3dCF7e'
